+++ utils/utils.hpp
@@ -23,7 +23,7 @@ double dSigmoid(double x);
 
 double ReLU(double x);
 
-int dReLU(double x);
+double dReLU(double x);
 
 double weightedSum(const std::vector<double> &weights, const std::vector<double> &inputs);
